refactor: Move `DeckSyncState` and `DeckMode` to domain package
*   Moves `DeckSyncState` from `ke.don.ski.navigation.deprecated` to a new file in `ke.don.ski.domain`.
*   Updates imports in `DeckWebImpl.wasmJs.kt` and `DeckWebImpl.js.kt` to reflect the new package locations for `DeckSyncState` and `DeckMode`.
*   Cleans up unused imports and redundant whitespace in `DeckWebImpl.js.kt` and `Deck.kt`.
