+++ composeApp/src/jsMain/kotlin/DeckWebImpl.js.kt
@@ -2,12 +2,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
-import ke.don.domain.Slide
 import ke.don.ski.Deck
+import ke.don.ski.domain.DeckMode
+import ke.don.ski.domain.DeckSyncState
 import ke.don.ski.navigation.DeckNavigator
-import ke.don.ski.navigation.deprecated.DeckSyncState
-import ke.don.ski.navigation.deprecated.rememberContainerState
-import ke.don.ski.presentation.deprecated.DeckMode
 import ke.don.ski.presentation.ui.skiPresentationSlides
 import kotlinx.browser.window
 import kotlinx.serialization.json.Json
